Fix cursor for avatar and user names in comments
When it was on an author row the cursor was shown as a pointer, even if
clicking on the author row itself does nothing. Also avatars used the
default cursor, even if clicking on them shows the contacts menu (if the
avatar is for a different user than the current one). Now the cursor is
shown as a pointer on avatar and user names of users different than the
current one.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>